testing: fix comp error
